+++ tests/testthat/helper-functions.R
@@ -1,3 +1,18 @@
+## Helper functions for package tests
+
+# Recursively round an object's numeric components
+#
+# Take an object and return it with numeric components rounded to a some
+# configurable number of significant figures. If the object is a language
+# object, return an identically structured object with rounding applied.
+# The main use case is to allow comparison of language objects with all.equal,
+# whose default method tests for equality of deparsed strings, to use a
+# numerical tolerance.
+#
+# @param x The object to recursively round.
+# @param digits The number of digits numeric components of x should be rounded to.
+# 
+# @return The passed object x with numeric components rounded.
 rec_round <-
 function(x, digits=8)
 {
@@ -17,4 +32,3 @@ function(x, digits=8)
     return(x)
   }
 }
-